move AthenzReqeustLog to server-common
