+++ libs/java/server_common/src/main/java/com/yahoo/athenz/common/server/log/jetty/AthenzRequestLog.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.yahoo.athenz.container.log;
+package com.yahoo.athenz.common.server.log.jetty;
 
 import java.io.IOException;
 import java.util.Locale;
